update application form and extract UI
